# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jobify API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Read Profile",
        "operationId": "read_profile_api_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/setup": {
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Put Setup",
        "operationId": "put_setup_api_v1_profile_setup_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchSetup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSetup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/cv": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Upload Cv",
        "description": "PDF or DOCX, up to 5 MB. Found skills come back as pending, for the user to confirm.",
        "operationId": "upload_cv_api_v1_profile_cv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_cv_api_v1_profile_cv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedCv"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/cv/items": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Add Item",
        "operationId": "add_item_api_v1_profile_cv_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CvItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CvItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/cv/items/{item_id}": {
      "patch": {
        "tags": [
          "profile"
        ],
        "summary": "Update Item",
        "operationId": "update_item_api_v1_profile_cv_items__item_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CvItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CvItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Delete Item",
        "operationId": "delete_item_api_v1_profile_cv_items__item_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_cv_api_v1_profile_cv_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_cv_api_v1_profile_cv_post"
      },
      "CvFileRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "contentType": {
            "type": "string",
            "title": "Contenttype"
          },
          "sizeBytes": {
            "type": "integer",
            "title": "Sizebytes"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Uploadedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fileName",
          "contentType",
          "sizeBytes",
          "uploadedAt"
        ],
        "title": "CvFileRead"
      },
      "CvItemCreate": {
        "properties": {
          "section": {
            "$ref": "#/components/schemas/CvSection"
          },
          "label": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Label"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "section",
          "label"
        ],
        "title": "CvItemCreate"
      },
      "CvItemRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "section": {
            "$ref": "#/components/schemas/CvSection"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "status": {
            "$ref": "#/components/schemas/CvItemStatus"
          },
          "origin": {
            "type": "string",
            "enum": [
              "cv",
              "manual"
            ],
            "title": "Origin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "section",
          "label",
          "detail",
          "confidence",
          "status",
          "origin"
        ],
        "title": "CvItemRead"
      },
      "CvItemStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed"
        ],
        "title": "CvItemStatus"
      },
      "CvItemUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CvItemStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "CvItemUpdate"
      },
      "CvSection": {
        "type": "string",
        "enum": [
          "skills",
          "experience",
          "education",
          "certifications"
        ],
        "title": "CvSection"
      },
...
      "EmploymentType": {
        "type": "string",
        "enum": [
          "full_time",
          "part_time",
          "contract",
          "internship"
        ],
        "title": "EmploymentType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParsedCv": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/CvFileRead"
          },
          "method": {
            "type": "string",
            "enum": [
              "basic",
              "ai"
            ],
            "title": "Method"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CvItemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "file",
          "method",
          "items",
          "notes"
        ],
        "title": "ParsedCv"
      },
      "ProfileRead": {
        "properties": {
          "setup": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchSetup"
              },
              {
                "type": "null"
              }
            ]
          },
          "cv": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CvFileRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CvItemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "completeness": {
            "type": "integer",
            "title": "Completeness"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing"
          }
        },
        "type": "object",
        "required": [
          "setup",
          "cv",
          "items",
          "completeness",
          "missing"
        ],
        "title": "ProfileRead"
      },
...
      "SalaryPeriod": {
        "type": "string",
        "enum": [
          "year",
          "month",
          "day",
          "hour"
        ],
        "title": "SalaryPeriod"
      },
      "SearchSetup": {
        "properties": {
          "targetRole": {
            "type": "string",
            "maxLength": 200,
            "title": "Targetrole",
            "default": ""
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "title": "Location",
            "default": ""
          },
          "salaryMin": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salarymin"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "salaryPeriod": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SalaryPeriod"
              },
              {
                "type": "null"
              }
            ]
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "employmentTypes": {
            "items": {
              "$ref": "#/components/schemas/EmploymentType"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Employmenttypes"
          },
          "workModes": {
            "items": {
              "$ref": "#/components/schemas/WorkMode"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Workmodes"
          }
        },
        "type": "object",
        "title": "SearchSetup"
      },
      "Seniority": {
        "type": "string",
        "enum": [
          "intern",
          "junior",
          "mid",
          "senior",
          "lead"
        ],
        "title": "Seniority"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkMode": {
        "type": "string",
        "enum": [
          "remote",
          "hybrid",
          "onsite"
        ],
        "title": "WorkMode"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}